Referencing: Replaced the return type of OperationMethod.source|targetDimensions() from int to Integer. This is an update for GEO-182 task on GeoAPI.

+++ modules/referencing/geotk-referencing/src/main/java/org/geotoolkit/referencing/operation/provider/EllipsoidToGeocentric.java
@@ -182,7 +182,7 @@ public class EllipsoidToGeocentric extends MathTransformProvider {
         final double semiMinor = doubleValue(SEMI_MINOR, values);
         final int    dimension = dimension(values);
         MathTransform transform = new GeocentricTransform(semiMajor, semiMinor, SI.METRE, dimension != 2);
-        if (dimension != sourceDimension) {
+        if (dimension != sourceDimension.intValue()) {
             transform = new MathTransformDecorator(transform, complement);
         }
         return transform;
